+++ NotificarePushKit/Sources/NotificarePush.swift
@@ -38,9 +38,6 @@ public protocol NotificarePush: AnyObject, NotificarePushUIApplicationDelegate, 
 
     func isNotificareNotification(_ userInfo: [AnyHashable: Any]) -> Bool
 
-    @available(*, deprecated, message: "Include the NotificareNotificationServiceExtensionKit and use NotificareNotificationServiceExtension.handleNotificationRequest() instead.")
-    func handleNotificationRequest(_ request: UNNotificationRequest, _ completion: @escaping NotificareCallback<UNNotificationContent>)
-
     @available(iOS 16.1, *)
     func registerLiveActivity(_ activityId: String, token: String, topics: [String], _ completion: @escaping NotificareCallback<Void>)